% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immediate_cost.R
\name{capacitated_random_assignment}
\alias{capacitated_random_assignment}
\title{Capacitated Random Assignment}
\usage{
capacitated_random_assignment(n, k, obj_, x, ...)
}
\arguments{
\item{n}{Total capacity.}

\item{k}{Number of sources.}

\item{obj_}{Objective coefficients.}

\item{x}{Capacity vector.}

\item{...}{Additional arguments.}
}
\value{
A list containing the assignment vector and objective value.
}
\description{
Generate a random assignment within capacity constraints.
}
