% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immediate_cost.R
\name{carrier_capacity}
\alias{carrier_capacity}
\title{carrier_capacity Function}
\usage{
carrier_capacity(env)
}
\arguments{
\item{env}{The environment variable.}
}
\value{
A list with constraints.
}
\description{
Calculate carrier capacity constraints.
}
