% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpmlt.R
\name{carrier_capacity_padded}
\alias{carrier_capacity_padded}
\title{Carrier Capacity with Padding}
\usage{
carrier_capacity_padded(env)
}
\arguments{
\item{env}{The environment list containing various parameters such as nI, nJ, nCS, nCO, and others.}
}
\value{
A list containing the matrix 'A' with added padding, right-hand side 'rhs', and 'sense'.
}
\description{
This function calculates the carrier capacity with added padding for the given environment.
}
