% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2STL.R
\name{CartesianProduct}
\alias{CartesianProduct}
\title{Compute Cartesian Indices}
\usage{
CartesianProduct(...)
}
\arguments{
\item{vectors}{A list of integer vectors.}
}
\value{
A matrix of integers representing the Cartesian indices.
}
\description{
This function takes a list of integer vectors and computes the Cartesian indices
using a C++ function.
}
\examples{
result <- computeCartesianIndices(c(1, 2), c(3, 4), c(5, 6))
print(result)
}
