% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{CartesianProductLB}
\alias{CartesianProductLB}
\title{Compute Cartesian Indices}
\usage{
CartesianProductLB(..., numThreads = 8)
}
\arguments{
\item{...}{Number of integer vectors.}

\item{numThreads}{Number of threads to use for parallel computation.}
}
\value{
A matrix of integers representing the Cartesian indices.
}
\description{
This function takes a list of integer vectors and computes the Cartesian indices
using a C++ function.
}
\examples{
result <- CartesianProductX(c(1, 2), c(3, 4), c(5, 6), numThreads = 4)
print(result)
}
