% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chunkup}
\alias{chunkup}
\title{Chunk Up a Sequence}
\usage{
chunkup(n, k)
}
\arguments{
\item{n}{Integer. The total length of the sequence to be divided.}

\item{k}{Integer. The number of chunks to divide the sequence into.}
}
\value{
A vector of integers representing the start of each chunk. The last element in the vector will always be \code{n + 1}, which represents the end of the last chunk.
}
\description{
This function divides a sequence of length \code{n} into \code{k} approximately equal-sized chunks.
}
\examples{
# Divide a sequence of 10 elements into 3 chunks
chunkup(10, 3)
# This will return something like: c(1, 4, 7, 11)

# Divide a sequence of 15 elements into 4 chunks
chunkup(15, 4)
# This will return something like: c(1, 4, 8, 12, 16)

}
