% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_graph_dt}
\alias{compute_graph_dt}
\title{Compute Graph Data Table}
\usage{
compute_graph_dt(env, S.I, S.J, allocation)
}
\arguments{
\item{env}{A list containing the environment variables including \code{L_}, \code{nL}, \code{tau}, \code{nvars}, and \code{L}.}

\item{S.I}{A vector representing the source nodes.}

\item{S.J}{A vector representing the destination nodes.}

\item{allocation}{A matrix representing the allocation data.}
}
\value{
A data.table containing the columns: \code{t} (time), \code{origin}, \code{destination}, and \code{assignment}.
The \code{assignment} column is the sum of allocations for each combination of time, origin, and destination.
}
\description{
This function computes a data.table representing the graph for a given environment,
source nodes (S.I), destination nodes (S.J), and allocation matrix.
}
\examples{
# Assuming `env`, `S.I`, `S.J`, and `allocation` are defined
graph_dt <- compute_graph_dt(env, S.I, S.J, allocation)

}
