% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{compute_stats}
\alias{compute_stats}
\title{Compute Statistics}
\usage{
compute_stats(v, N)
}
\arguments{
\item{v}{The data.}

\item{N}{The number of observations.}
}
\description{
Computes statistics for a given set of data.
}
