% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{consolidate_idx}
\alias{consolidate_idx}
\title{Consolidate indices from lists using data.table::CJ}
\usage{
consolidate_idx(lists)
}
\arguments{
\item{lists}{A list of indices.}
}
\value{
A matrix where each row represents a unique combination of indices.
}
\description{
This function takes a list of indices and consolidates them into a matrix where each row contains a unique combination of indices using data.table::CJ function.
}
\examples{
lists <- list(c(1, 2), c(3, 4), c(5, 6))
consolidate_idx(lists)
}
