% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\name{consolidate_idx__}
\alias{consolidate_idx__}
\title{Consolidate indices from lists}
\usage{
consolidate_idx__(lists)
}
\arguments{
\item{lists}{A list of indices.}
}
\value{
A matrix where each row represents a unique combination of indices.
}
\description{
This function takes a list of indices and consolidates them into a matrix where each row contains a unique combination of indices.
}
\examples{
lists <- list(c(1, 2), c(3, 4), c(5, 6))
consolidate_idx__(lists)
}
