% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2STL.R
\name{createAndPassList}
\alias{createAndPassList}
\title{Process List of Vectors}
\usage{
createAndPassList(list_of_vectors)
}
\description{
This function creates a list of numeric vectors and passes it to a C++ function
to be processed.
}
\examples{
createAndPassList()
}
