% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portf_eval.R
\name{eval_portfolio}
\alias{eval_portfolio}
\title{Evaluating a portfolio contract with state-dependent constraints}
\usage{
eval_portfolio(env, model, A_, S0, Q, D, ...)
}
\arguments{
\item{env}{The environment containing the parameters of the portfolio contract.}

\item{model}{The LP model created by \code{create_model(env)}.}

\item{A_}{A vector containing the values of the assets.}

\item{S0}{Initial state vector (1 x (nI + nJ) matrix).}

\item{Q}{Vector of supply realisations (length tau * nI).}

\item{D}{Vector of demand realisations (length tau * nJ).}

\item{...}{Additional arguments (not used).}
}
\value{
A list containing the states, cost, and allocation of the portfolio contract.
}
\description{
This function evaluates a portfolio contract considering state-dependent constraints.
}
