% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portf_eval.R
\name{eval_stateless_portfolio}
\alias{eval_stateless_portfolio}
\title{Evaluating a portfolio contract without state-dependent constraints}
\usage{
eval_stateless_portfolio(env, A_, plot = FALSE)
}
\arguments{
\item{env}{The environment containing the parameters of the portfolio contract.}

\item{A_}{A vector containing the values of the assets.}

\item{plot}{Logical indicating whether to plot the cost.}
}
\value{
A list containing the cost and allocation of the portfolio contract.
}
\description{
This function evaluates a portfolio contract without state-dependent constraints.
}
\details{
This function evaluates a portfolio contract without considering state-dependent constraints.
It computes the cost and allocation for each asset value provided in the vector \code{A_}. If \code{plot} is \code{TRUE}, it plots the cost.
}
