% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cssap_gen.R
\name{generate_bids}
\alias{generate_bids}
\title{Generate Bids}
\usage{
generate_bids(nB, L)
}
\arguments{
\item{nB}{The number of bids.}

\item{L}{A matrix of lanes.}
}
\value{
A list of bids.
}
\description{
Generates multiple bids.
}
