% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cssap_gen.R
\name{generate_cssap}
\alias{generate_cssap}
\title{Generate CSSAP}
\usage{
generate_cssap(env, tau = 12L, ...)
}
\arguments{
\item{env}{The environment variable.}

\item{tau}{The time horizon. Defaults to 12L.}

\item{...}{Additional arguments.}
}
\description{
Generates a CSSAP (Carrier Scheduling and Shipment Assignment Problem).
}
