% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generatePositiveDefiniteMatrix}
\alias{generatePositiveDefiniteMatrix}
\title{Generate a Positive Definite Matrix}
\usage{
generatePositiveDefiniteMatrix(p)
}
\arguments{
\item{p}{An integer specifying the dimension of the matrix (number of rows and columns).}
}
\value{
A p x p symmetric positive definite matrix.
}
\description{
This function generates a random symmetric positive definite matrix of a given size.
}
\examples{
# Generate a 3x3 positive definite matrix
matrix <- generatePositiveDefiniteMatrix(3)
print(matrix)
}
