% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\name{get_adjustment_weights}
\alias{get_adjustment_weights}
\title{Generate Adjustment Weights}
\usage{
get_adjustment_weights(env)
}
\arguments{
\item{env}{An environment containing the variables \code{nSI}, \code{nI}, \code{nJ}, and \code{nSJ}.
\itemize{
\item \code{nSI} is an integer representing a scaling factor for the first group.
\item \code{nI} is an integer indicating the number of elements in the first group.
\item \code{nJ} is an integer representing the number of elements in the second group.
\item \code{nSJ} is an integer representing a scaling factor for the second group.
}}
}
\value{
A numeric vector containing the adjustment weights.
}
\description{
This function generates a vector of adjustment weights based on the input environment \code{env}.
The weights are computed using powers of \code{nSI} and \code{nSJ} (if \code{nJ} is greater than 1).
}
