% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cssap_gen.R
\name{get_from_routes}
\alias{get_from_routes}
\title{Get From Routes}
\usage{
get_from_routes(env, ...)
}
\arguments{
\item{env}{The environment variable.}
}
\description{
Retrieves the routes originating from different sources.
}
