% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cssap_gen.R
\name{get_lanes}
\alias{get_lanes}
\title{Get Lanes}
\usage{
get_lanes(I, J)
}
\arguments{
\item{I}{A vector representing the sources.}

\item{J}{A vector representing the destinations.}
}
\value{
A matrix of lanes.
}
\description{
Generates a matrix of lanes.
}
