% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_Qdx_fixed_exit}
\alias{get_Qdx_fixed_exit}
\title{Action-Value Surface for Fixed Exit State}
\usage{
get_Qdx_fixed_exit(env, exit_state, ...)
}
\arguments{
\item{env}{List. The environment object containing the state support, action indices, and other relevant parameters.}

\item{exit_state}{Numeric vector. The fixed exit state used to calculate the action-value surface.}

\item{...}{Additional arguments passed to the function.}
}
\value{
An integer vector representing the action-value surface indices.
}
\description{
This function computes the action-value surface given a fixed exit state in a specified environment.
It calculates the state-action index values based on the state support and the environment configuration.
}
\examples{
env <- list(nI = 2, stateSupport = c(1, 2), R = 1, nAdx = 3, stateKeys = c(1, 10))
exit_state <- c(1, 2)
get_Qdx_fixed_exit(env, exit_state)

}
