% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\name{get_scenario_space}
\alias{get_scenario_space}
\title{Get scenario space}
\usage{
get_scenario_space(env, nI, nJ, Q, D, W, hmg = TRUE)
}
\arguments{
\item{Q}{Q parameter}

\item{D}{D parameter}

\item{W}{W parameter}

\item{hmg}{Boolean indicating whether to use hmg or not (default: TRUE)}
}
\description{
Get scenario space
}
