% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\name{get_state_indices}
\alias{get_state_indices}
\title{Endogenous state vector}
\usage{
get_state_indices(env, nI, nJ, max_, incr_)
}
\arguments{
\item{env}{Environment to host dynamic programming data}

\item{nI}{Number of origin hubs}

\item{nJ}{Number of destination hubs}

\item{max_}{Maximum value}

\item{incr_}{Increment value}
}
\description{
Endogenous state vector
}
