% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{get_to_routes}
\alias{get_to_routes}
\title{Get To Routes}
\usage{
get_to_routes(env, ...)
}
\arguments{
\item{env}{The environment variable.}
}
\description{
Retrieves the routes leading to different destinations.
}
