% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immediate_cost.R
\name{h.t}
\alias{h.t}
\title{h.t Function}
\usage{
h.t(env, SIt, SJt, alpha = NULL)
}
\arguments{
\item{env}{The environment variable.}

\item{SIt}{Some variable.}

\item{SJt}{Another variable.}

\item{alpha}{A coefficient vector. Defaults to NULL.}
}
\value{
The calculated cost.
}
\description{
Calculate some cost based on inputs.
}
