% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immediate_cost.R
\name{heuristic_assignment}
\alias{heuristic_assignment}
\title{Heuristic Assignment Function}
\usage{
heuristic_assignment(model, obj_, rhs_, k, edx, ...)
}
\arguments{
\item{model}{A list containing the model matrix \code{A} and \code{sense} vector.}

\item{obj_}{A numeric vector representing the objective coefficients.}

\item{rhs_}{A numeric vector representing the right-hand side constraints.}

\item{k}{An integer representing the number of variables.}

\item{edx}{An integer index to specify a subset of constraints to be excluded.}

\item{...}{Additional arguments (not used).}
}
\value{
A list containing:
\describe{
\item{x}{A numeric vector representing the assignment solution.}
\item{objval}{A numeric value representing the objective value of the solution.}
\item{status}{A character string representing the status of the solution.}
}
}
\description{
This function performs a heuristic assignment given a model and constraints.
}
\examples{
model <- list(
  A = matrix(c(1, 2, 3, 4, 5, 6), nrow = 2, byrow = TRUE),
  sense = c("<=", "<=")
)
obj_ <- c(1, 2, 3)
rhs_ <- c(7, 8)
k <- 3
edx <- 1
heuristic_assignment(model, obj_, rhs_, k, edx)

}
