% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cssap_gen.R
\name{init_env}
\alias{init_env}
\title{Initialize Environment}
\usage{
init_env(env, ...)
}
\arguments{
\item{env}{The environment variable.}

\item{...}{Additional arguments.}
}
\description{
Sets up the drayage environment: storage limits, carrier information, cost
parameters, capacity limits, and route indexing.
}
