% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cssap_gen.R
\name{initialize}
\alias{initialize}
\title{Initialize}
\usage{
initialize(env, ...)
}
\arguments{
\item{env}{The environment variable.}

\item{...}{Additional arguments.}
}
\description{
Initializes the environment.
}
