% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cssap_gen.R
\name{input_configuration}
\alias{input_configuration}
\title{Input Configuration}
\usage{
input_configuration(env, ...)
}
\arguments{
\item{env}{The environment variable.}

\item{...}{Additional arguments.}
}
\description{
Configures the input environment.
}
