% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpmlt.R
\name{multiperiod_expansion}
\alias{multiperiod_expansion}
\title{Multiperiod Expansion}
\usage{
multiperiod_expansion(env, A, obj, rhs, sns)
}
\arguments{
\item{env}{The environment list containing various parameters such as nI, nJ, nvars, tau, alpha, and others.}

\item{A}{The constraint matrix.}

\item{obj}{The objective function coefficients.}

\item{rhs}{The right-hand side vector.}

\item{sns}{The sense vector.}
}
\value{
A list containing the expanded objective function 'obj', constraint matrix 'A', right-hand side 'rhs', and 'sense'.
}
\description{
This function expands the given constraints and objective function for multiple periods.
}
