% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immediate_cost.R
\name{optimal_assignment}
\alias{optimal_assignment}
\title{Optimal Single Stage Shipment Assignment}
\usage{
optimal_assignment(model, obj_, rhs_, params = list(OutputFlag = 0L), ...)
}
\arguments{
\item{model}{The optimization model.}

\item{obj_}{The objective function coefficients.}

\item{rhs_}{The right-hand side of constraints.}

\item{params}{A list of parameters for the optimization solver. Defaults to list(OutputFlag = 0L).}

\item{...}{Additional arguments.}
}
\value{
The result of the optimization.
}
\description{
Solves for the optimal assignment of shipments.
}
