% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{optimizeModelFromJSON}
\alias{optimizeModelFromJSON}
\title{Optimize Model from JSON file}
\usage{
optimizeModelFromJSON(jsonFile)
}
\arguments{
\item{json_file}{Path to the JSON file.}
}
\description{
This function takes a JSON file and optimizes the model using a C++ function.
}
\examples{
result <- optimizeModelFromJSON("path/to/json/file.json")
print(result$objval)
print(result$x)
}
