% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immediate_cost.R
\name{order_quantity}
\alias{order_quantity}
\title{order_quantity Function}
\usage{
order_quantity(env, ...)
}
\arguments{
\item{env}{The environment variable.}
}
\value{
A list with constraints.
}
\description{
Calculate order quantity constraints.
}
