% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{passRmatrixToCPP}
\alias{passRmatrixToCPP}
\title{Process R matrix object}
\usage{
passRmatrixToCPP(R_matrix)
}
\description{
This function takes an R matrix object and passes it to a C++ function to be
processed.
}
\examples{
passRmatrixToCPP()
}
