% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot2x2instance}
\alias{plot2x2instance}
\title{Plot 2x2 Instance}
\usage{
plot2x2instance(env, graph.dt, S.I, S.J, Q, D)
}
\arguments{
\item{env}{A list containing the environment variables including \code{tau}, \code{nI}, and \code{nJ}.}

\item{graph.dt}{A data.table containing the graph data with columns \code{t}, \code{origin}, \code{destination}, and \code{assignment}.}

\item{S.I}{A vector representing the source nodes.}

\item{S.J}{A vector representing the destination nodes.}

\item{Q}{A vector representing some quantities related to the nodes.}

\item{D}{A vector representing some other quantities related to the nodes.}
}
\description{
This function plots a series of graphs representing the state of the environment at different times.
}
\examples{
# Assuming `env`, `graph.dt`, `S.I`, `S.J`, `Q`, and `D` are defined
plot2x2instance(env, graph.dt, S.I, S.J, Q, D)

}
