% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plotVsurface}
\alias{plotVsurface}
\title{Plot State Value Surface}
\usage{
plotVsurface(VT, t, theta = 30, phi = 30)
}
\arguments{
\item{VT}{Numeric matrix. The value matrix to plot.}

\item{t}{Integer. The time index for which the value surface is plotted.}

\item{theta}{Numeric. The angle defining the viewing direction. Default is \code{30}.}

\item{phi}{Numeric. The angle defining the viewing direction. Default is \code{30}.}
}
\value{
A 3D plot of the state value surface.
}
\description{
This function creates a 3D plot of the state value surface at a given time \code{t}.
The z-values are normalized and mapped to colors using a gradient palette.
}
\examples{
# Assuming 'VT' is a predefined value matrix and 't' is a valid time index
plotVsurface(VT, t = 5)

}
