% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immediate_cost.R
\name{positivity_}
\alias{positivity_}
\title{positivity_ Function}
\usage{
positivity_(env, ...)
}
\arguments{
\item{env}{The environment variable.}
}
\value{
A list with constraints.
}
\description{
Calculate positivity constraints.
}
