% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpmlt.R
\name{post_hoc_simulation}
\alias{post_hoc_simulation}
\title{Post-Hoc Simulation}
\usage{
post_hoc_simulation(env, x)
}
\arguments{
\item{env}{The environment list containing various parameters such as tau, nI, nJ, nvars, alpha, R, I_, J_, and others.}

\item{x}{A numeric vector representing the decision variables.}
}
\value{
A list containing the simulated inventory levels 'S.I', job status 'S.J', and 'allocation'.
}
\description{
This function simulates the results post hoc for the given environment and decision variables.
}
