% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cssap_gen.R
\name{print_info}
\alias{print_info}
\title{Print Info}
\usage{
print_info(env, ...)
}
\arguments{
\item{env}{The environment variable.}

\item{...}{Additional arguments.}
}
\description{
Prints information about the auction.
}
