% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immediate_cost.R
\name{random_assignment}
\alias{random_assignment}
\title{Random Single Stage Shipment Assignment}
\usage{
random_assignment(n, k, obj_, ...)
}
\arguments{
\item{n}{The number of shipments.}

\item{k}{The number of carriers.}

\item{obj_}{The objective function coefficients.}

\item{...}{Additional arguments.}
}
\value{
A list containing the assignment and objective value.
}
\description{
Generates a random assignment of shipments.
}
