% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{random_volume}
\alias{random_volume}
\title{Random Volume Selection}
\usage{
random_volume(capacity, limit, ...)
}
\arguments{
\item{capacity}{A numeric vector representing the available capacities.}

\item{limit}{A numeric value representing the upper limit for the selected volume.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A numeric value representing the selected volume, which is the minimum of a random sample from the capacity and the provided limit.
}
\description{
Selects a random volume based on the given capacity and limits.
}
