% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rev_cols}
\alias{rev_cols}
\title{Reverse the columns of a matrix}
\usage{
rev_cols(X)
}
\arguments{
\item{X}{A matrix.}
}
\value{
A matrix with columns in reversed order.
}
\description{
This function reverses the order of columns in a matrix.
}
\examples{
mat <- matrix(1:9, ncol = 3)
rev_cols(mat)
}
