% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\name{run_scenarios}
\alias{run_scenarios}
\title{Run scenarios}
\usage{
run_scenarios(env, t, start, end, scenaria, weights, ...)
}
\arguments{
\item{env}{Environment object containing relevant data}

\item{t}{Value representing time}

\item{start}{Start index}

\item{end}{End index}

\item{scenaria}{A vector of scenarios}

\item{weights}{Numeric vector of weights}
}
\value{
A matrix representing transit
}
\description{
Run scenarios
}
