% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sample_exogenous_state}
\alias{sample_exogenous_state}
\title{Sample Exogenous State}
\usage{
sample_exogenous_state(func, params, ...)
}
\arguments{
\item{func}{The function for sampling.}

\item{params}{The parameters for the sampling function.}
}
\description{
Samples the exogenous state based on a given function and parameters.
}
