% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cssap_gen.R
\name{simulate_auction}
\alias{simulate_auction}
\title{Simulate Auction}
\usage{
simulate_auction(env, ...)
}
\arguments{
\item{env}{The environment variable.}

\item{...}{Additional arguments.}
}
\description{
Simulates an auction.
}
