% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_system}
\alias{simulate_system}
\title{Simulate System}
\usage{
simulate_system(env, policy, args, exog, RHSdx, correction = FALSE, ...)
}
\arguments{
\item{env}{The environment variable.}

\item{policy}{The policy for simulation.}

\item{args}{Arguments for policy functions.}

\item{exog}{Exogenous states for simulation.}
}
\description{
Simulates the system based on policies, exogenous states, and environment parameters.
}
