% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cssap_gen.R
\name{simulate_winners}
\alias{simulate_winners}
\title{Simulate Winners}
\usage{
simulate_winners(nB, nCS)
}
\arguments{
\item{nB}{The number of bids.}

\item{nCS}{The number of strategic carriers.}
}
\value{
A list of winners.
}
\description{
Simulates the winners of bids.
}
