% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\name{stateIdx}
\alias{stateIdx}
\title{All possible allocations of n containers to k routes}
\usage{
stateIdx(env, Si, Sj, max.S, weights, ...)
}
\arguments{
\item{env}{Environment object containing relevant data}

\item{Si}{Value representing some state information}

\item{Sj}{Value representing some state information}

\item{max.S}{Maximum value for a state}

\item{weights}{Numeric vector of weights}
}
\value{
A numeric vector representing state index
}
\description{
All possible allocations of n containers to k routes
}
