% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immediate_cost.R
\name{storage_limit}
\alias{storage_limit}
\title{storage_limit Function}
\usage{
storage_limit(env, ...)
}
\arguments{
\item{env}{The environment variable.}
}
\value{
A list with constraints.
}
\description{
Calculate storage constraints.
}
