% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpmlt.R
\name{storage_limits}
\alias{storage_limits}
\title{Storage Limits}
\usage{
storage_limits(env, q)
}
\arguments{
\item{env}{The environment list containing various parameters such as nI, nJ, nvars, and others.}

\item{q}{A numeric vector representing the initial conditions.}
}
\value{
A list containing the matrix 'A', right-hand side 'rhs', and 'sense'.
}
\description{
This function constructs the storage limit constraints for the given environment.
}
