% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpmlt.R
\name{transition_logic}
\alias{transition_logic}
\title{Transition Logic}
\usage{
transition_logic(env, q, d)
}
\arguments{
\item{env}{The environment list containing various parameters such as nI, nJ, nvars, and others.}

\item{q}{A numeric vector representing the initial conditions.}

\item{d}{A numeric vector representing the demands.}
}
\value{
A list containing the matrix 'A', right-hand side 'rhs', and 'sense'.
}
\description{
This function constructs the transition logic matrix and its components for the given environment.
}
