% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfa.R
\name{vfa_init}
\alias{vfa_init}
\title{Initialise a zero-valued separable VFA parameter object}
\usage{
vfa_init(env)
}
\arguments{
\item{env}{Environment produced by \code{dp_config}.}
}
\value{
A named list with elements \code{orig} and \code{dest}.
}
\description{
Initialise a zero-valued separable VFA parameter object
}
