% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfa.R
\name{vfa_project_concavity}
\alias{vfa_project_concavity}
\title{Apply concavity projection to all 1-D functions in a theta object}
\usage{
vfa_project_concavity(theta)
}
\arguments{
\item{theta}{Parameter object from \code{vfa_init} or \code{vfa_fit}.}
}
\value{
Updated \code{theta} with concave \code{orig} and \code{dest}
columns.
}
\description{
Apply concavity projection to all 1-D functions in a theta object
}
