% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\name{compute_environment}
\alias{compute_environment}
\title{Run scenarios}
\usage{
compute_environment(env, env.dp, t, start, end, scenaria, weights, FUN, ...)
}
\arguments{
\item{env}{Environment object containing relevant data}

\item{env.dp}{Environment object containing dynamic programming data}

\item{t}{Value representing time}

\item{start}{Start index}

\item{end}{End index}

\item{scenaria}{A vector of scenario indexes}

\item{weights}{Numeric vector of weights}

\item{FUN}{The assignment algorithm}
}
\value{
A matrix representing transit
}
\description{
Run scenarios
}
