% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immediate_cost.R
\name{create_model}
\alias{create_model}
\title{create_model Function}
\usage{
create_model(env, Idx = seq(4L), ...)
}
\arguments{
\item{env}{The environment variable.}

\item{...}{Additional arguments.}
}
\value{
A model.
}
\description{
Create a model with specified constraints.
}
